Moving IdentifierGov copy map back to parent file, updating snapshots
